refactor : 테스트 하기 쉬운 코드로 리팩토링
